rename auto_serialization.h/.cpp
