+++ cpp/memilio/CMakeLists.txt
@@ -26,8 +26,8 @@ add_library(memilio
     compartments/simulation.h
     compartments/flow_simulation.h
     compartments/parameter_studies.h
-    io/auto_serialize.h
-    io/auto_serialize.cpp
+    io/default_serialize.h
+    io/default_serialize.cpp
     io/io.h
     io/io.cpp
     io/hdf5_cpp.h
