+++ cpp/memilio/io/default_serialize.cpp
@@ -17,4 +17,4 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-#include "memilio/io/auto_serialize.h"
+#include "memilio/io/default_serialize.h"
