+++ cpp/memilio/utils/random_number_generator.h
@@ -21,7 +21,7 @@
 #ifndef MIO_RANDOM_NUMBER_GENERATOR_H
 #define MIO_RANDOM_NUMBER_GENERATOR_H
 
-#include "memilio/io/auto_serialize.h"
+#include "memilio/io/default_serialize.h"
 #include "memilio/utils/compiler_diagnostics.h"
 #include "memilio/utils/logging.h"
 #include "memilio/utils/miompi.h"
